removed old tests and formatted unit tests for t-route routing
